+++ packages/core/modal-dialog/src/components/__tests__/visual-regression/modalSnapshotTest.ts
@@ -6,8 +6,7 @@ import {
 const openModalBtn = "[type='button']";
 const modalDialog = "[role='dialog']";
 
-// TODO: https://ecosystem.atlassian.net/browse/AK-5842
-describe.skip('Snapshot Test', () => {
+describe('Snapshot Test', () => {
   it('Basic example should match production example', async () => {
     const url = getExampleUrl(
       'core',
@@ -22,8 +21,9 @@ describe.skip('Snapshot Test', () => {
     await page.goto(url);
     await page.waitForSelector(openModalBtn);
     await page.click(openModalBtn);
-    await page.waitFor(modalDialog);
-
+    await page.waitForSelector(modalDialog);
+    // We need to wait for the animation to finish.
+    await page.waitFor(1000);
     const image = await takeElementScreenShot(page, modalDialog);
     expect(image).toMatchProdImageSnapshot();
   });