+++ packages/core/onboarding/src/components/__tests__/visual-regression/onBoardingSnapshotTest.js
@@ -8,8 +8,7 @@ import {
 const openModalBtn = "[type='button']";
 const modalDialog = "[role='dialog']";
 
-// TODO: https://ecosystem.atlassian.net/browse/AK-5842
-describe.skip('Snapshot Test', () => {
+describe('Snapshot Test', () => {
   it('Spotlight different-spotlights should match production example', async () => {
     const url = getExampleUrl(
       'core',
@@ -21,6 +20,7 @@ describe.skip('Snapshot Test', () => {
     //$FlowFixMe
     expect(image).toMatchProdImageSnapshot();
   });
+
   it('Modal Basic example should match production example', async () => {
     const url = getExampleUrl(
       'core',
@@ -33,7 +33,9 @@ describe.skip('Snapshot Test', () => {
     await page.goto(url);
     await page.waitForSelector(openModalBtn);
     await page.click(openModalBtn);
-    await page.waitFor(modalDialog);
+    await page.waitForSelector(modalDialog);
+    // We need to wait for the animation to finish.
+    await page.waitFor(1000);
 
     const image = await takeElementScreenShot(page, modalDialog);
     //$FlowFixMe