+++ packages/core/select/src/components/__tests__/visual-regression/selectSnapshotTest.js
@@ -7,8 +7,7 @@ import {
 const openModalBtn = "[type='button']";
 const modalDialog = "[role='dialog']";
 
-// TODO: https://ecosystem.atlassian.net/browse/AK-5843
-describe.skip('Snapshot Test', () => {
+describe('Snapshot Test', () => {
   it('Select in a modal dialog example should match production example', async () => {
     const url = getExampleUrl(
       'core',
@@ -21,7 +20,9 @@ describe.skip('Snapshot Test', () => {
     await page.goto(url);
     await page.waitForSelector(openModalBtn);
     await page.click(openModalBtn);
-    await page.waitFor(modalDialog);
+    await page.waitForSelector(modalDialog);
+    // We need to wait for the animation to finish.
+    await page.waitFor(1000);
 
     const image = await takeElementScreenShot(page, modalDialog);
     //$FlowFixMe