+++ packages/editor/editor-core/src/__tests__/unit/plugins/panel/nodeviews/panel.ts
@@ -2,16 +2,10 @@ import { p, panel, defaultSchema } from '@atlaskit/editor-test-helpers';
 import { panelNodeView } from '../../../../../plugins/panel/nodeviews/panel';
 
 describe('Panel - NodeView', () => {
-  const portalProviderAPI = { render() {}, remove() {} } as any;
-
   it('should render a contentDOM of `div` inside `div[data-panel-type]`', () => {
     const node = panel()(p('this is the decision'))(defaultSchema);
 
-    const nodeView = panelNodeView(portalProviderAPI)(
-      node,
-      null as any,
-      () => -1,
-    );
+    const nodeView = panelNodeView()(node);
 
     const contentDOM = nodeView!.contentDOM as HTMLElement;