+++ packages/editor/editor-core/src/index.ts
@@ -60,7 +60,7 @@ export {
   toggleStrong,
 } from './plugins/text-formatting/commands/text-formatting';
 export { insertBlockType, setBlockType } from './plugins/block-type/commands';
-export { createTable } from './plugins/table/actions';
+export { createTable } from './plugins/table/commands';
 export { insertTaskDecision } from './plugins/tasks-and-decisions/commands';
 export { EventDispatcher } from './event-dispatcher';
 export {