+++ packages/media/media-picker/src/components/types.ts
@@ -18,16 +18,6 @@ export interface LocalUploadComponent<
   setUploadParams(uploadParams: UploadParams): void;
 }
 
-export type BinaryConfig = LocalUploadConfig;
-
-export interface BinaryUploaderConstructor {
-  new (context: Context, config: BinaryConfig): BinaryUploader;
-}
-
-export interface BinaryUploader extends LocalUploadComponent {
-  upload(base64: string, name: string): void;
-}
-
 export interface BrowserConfig extends LocalUploadConfig {
   readonly multiple?: boolean;
   readonly fileExtensions?: Array<string>;