+++ packages/media/media-picker/src/popup/components/app.tsx
@@ -12,7 +12,6 @@ import {
 
 import { ServiceName, State } from '../domain';
 
-import { BinaryUploaderImpl as MpBinary } from '../../components/binary';
 import { BrowserImpl as MpBrowser } from '../../components/browser';
 import { DropzoneImpl as MpDropzone } from '../../components/dropzone';
 import { ClipboardImpl as MpClipboard } from '../../components/clipboard';
@@ -56,6 +55,7 @@ import {
   DropzoneDragEnterEventPayload,
   DropzoneDragLeaveEventPayload,
 } from '../../components/types';
+import { LocalUploadComponent } from '../../components/localUpload';
 
 export interface AppStateProps {
   readonly selectedServiceName: ServiceName;
@@ -104,8 +104,8 @@ export interface AppState {
 export class App extends Component<AppProps, AppState> {
   private readonly mpBrowser: MpBrowser;
   private readonly mpDropzone: MpDropzone;
-  private readonly mpBinary: MpBinary;
   private readonly mpClipboard: MpClipboard;
+  private readonly localUploader: LocalUploadComponent;
 
   constructor(props: AppProps) {
     super(props);
@@ -135,11 +135,24 @@ export class App extends Component<AppProps, AppState> {
       cacheSize: tenantContext.config.cacheSize,
     });
 
+    this.localUploader = new LocalUploadComponent(context, {
+      uploadParams: tenantUploadParams,
+      shouldCopyFileToRecents: false,
+    });
+
+    this.localUploader.on('uploads-start', onUploadsStart);
+    this.localUploader.on('upload-preview-update', onUploadPreviewUpdate);
+    this.localUploader.on('upload-status-update', onUploadStatusUpdate);
+    this.localUploader.on('upload-processing', onUploadProcessing);
+    this.localUploader.on('upload-end', onUploadEnd);
+    this.localUploader.on('upload-error', onUploadError);
+
     this.mpBrowser = new MpBrowser(context, {
       uploadParams: tenantUploadParams,
       shouldCopyFileToRecents: false,
       multiple: true,
     });
+
     this.mpBrowser.on('uploads-start', onUploadsStart);
     this.mpBrowser.on('upload-preview-update', onUploadPreviewUpdate);
     this.mpBrowser.on('upload-status-update', onUploadStatusUpdate);
@@ -161,17 +174,6 @@ export class App extends Component<AppProps, AppState> {
     this.mpDropzone.on('upload-end', onUploadEnd);
     this.mpDropzone.on('upload-error', onUploadError);
 
-    this.mpBinary = new MpBinary(context, {
-      uploadParams: tenantUploadParams,
-      shouldCopyFileToRecents: false,
-    });
-    this.mpBinary.on('uploads-start', onUploadsStart);
-    this.mpBinary.on('upload-preview-update', onUploadPreviewUpdate);
-    this.mpBinary.on('upload-status-update', onUploadStatusUpdate);
-    this.mpBinary.on('upload-processing', onUploadProcessing);
-    this.mpBinary.on('upload-end', onUploadEnd);
-    this.mpBinary.on('upload-error', onUploadError);
-
     this.mpClipboard = new MpClipboard(context, {
       uploadParams: tenantUploadParams,
       shouldCopyFileToRecents: false,
@@ -188,7 +190,7 @@ export class App extends Component<AppProps, AppState> {
       onCancelUpload: uploadId => {
         this.mpBrowser.cancel(uploadId);
         this.mpDropzone.cancel(uploadId);
-        this.mpBinary.cancel(uploadId);
+        this.localUploader.cancel(uploadId);
       },
     });
   }
@@ -253,7 +255,7 @@ export class App extends Component<AppProps, AppState> {
                     <Footer />
                   </ViewWrapper>
                   <Dropzone isActive={isDropzoneActive} />
-                  <MainEditorView binaryUploader={this.mpBinary} />
+                  <MainEditorView localUploader={this.localUploader} />
                 </MediaPickerPopupWrapper>
               </PassContext>
             </ModalDialog>