+++ packages/media/media-picker/src/popup/components/views/editor/__tests__/unit/mainEditorView.spec.tsx
@@ -13,7 +13,6 @@ import {
 } from '../../mainEditorView';
 import { ErrorView } from '../../errorView/errorView';
 import { SpinnerView } from '../../spinnerView/spinnerView';
-import { BinaryUploader } from '../../../../../..';
 
 describe('MainEditorView', () => {
   class FakeEditorView extends Component<{}, {}> {
@@ -24,16 +23,15 @@ describe('MainEditorView', () => {
 
   const setup = (props?: Partial<MainEditorViewStateProps>) => {
     const editorLoaderPromise = Promise.resolve(FakeEditorView);
-    const upload: BinaryUploader['upload'] = jest.fn();
-    const binaryUploader = {
-      upload,
-    } as BinaryUploader;
     const onCloseEditor: MainEditorViewDispatchProps['onCloseEditor'] = jest.fn();
     const onShowEditorError: MainEditorViewDispatchProps['onShowEditorError'] = jest.fn();
     const onDeselectFile: MainEditorViewDispatchProps['onDeselectFile'] = jest.fn();
+    const localUploader: any = {
+      addFiles: jest.fn(),
+    };
     const mainView = shallow(
       <MainEditorView
-        binaryUploader={binaryUploader}
+        localUploader={localUploader}
         editorData={{}}
         onCloseEditor={onCloseEditor}
         onShowEditorError={onShowEditorError}
@@ -45,8 +43,8 @@ describe('MainEditorView', () => {
     return {
       mainView,
       editorLoaderPromise,
-      binaryUploader,
       onCloseEditor,
+      localUploader,
     };
   };
 
@@ -78,7 +76,7 @@ describe('MainEditorView', () => {
     expectToEqual(mainView.find(EditorView).props().imageUrl, 'some-image-url');
   });
   it('should upload an image and call onCloseEditor when editor viewer calls onSave', () => {
-    const { mainView, binaryUploader, onCloseEditor } = setup({
+    const { mainView, onCloseEditor, localUploader } = setup({
       editorData: {
         imageUrl: 'some-image-url',
         originalFile: {
@@ -87,13 +85,16 @@ describe('MainEditorView', () => {
         },
       },
     });
+
+    const instance: any = mainView.instance();
+    instance.urltoFile = jest.fn().mockReturnValue('some-image-string');
+
     mainView
       .find(EditorView)
       .props()
       .onSave('some-image-string', { width: 200, height: 100 });
-    expectFunctionToHaveBeenCalledWith(binaryUploader.upload, [
-      'some-image-string',
-      'some-file-name',
+    expectFunctionToHaveBeenCalledWith(localUploader.addFiles, [
+      ['some-image-string'],
     ]);
     expectFunctionToHaveBeenCalledWith(onCloseEditor, ['Save']);
   });