+++ packages/editor/conversation/CHANGELOG.md
@@ -1,5 +1,10 @@
 # @atlaskit/conversation
 
+## 15.0.1
+- Updated dependencies [154372926b](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/154372926b):
+  - @atlaskit/comment@8.0.1
+  - @atlaskit/editor-core@111.0.0
+
 ## 15.0.0
 - [major] [7c17b35107](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/7c17b35107):