+++ packages/editor/editor-wikimarkup-transformer/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/editor-wikimarkup-transformer
 
+## 4.1.1
+- Updated dependencies [154372926b](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/154372926b):
+  - @atlaskit/editor-json-transformer@6.0.1
+  - @atlaskit/editor-test-helpers@9.1.2
+  - @atlaskit/editor-core@111.0.0
+
 ## 4.1.0
 - [minor] [5a49043dac](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/5a49043dac):