+++ packages/help/help-article/CHANGELOG.md
@@ -1,5 +1,10 @@
 # @atlaskit/help-article
 
+## 0.4.0
+- [minor] [05460c129b](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/05460c129b):
+
+  - Added prop titleLinkUrl to make the title of the article a link
+
 ## 0.3.1
 - [patch] [d3a2a15890](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/d3a2a15890):